+++ NEWS.md
@@ -1,3 +1,5 @@
+# hpa.spatial (development version)
+
 # hpa.spatial 0.2.6
 
 * make `create_child_geo()` which adapts an `sf` polygon (`child_geo`) object by